+++ tests/test_transaction.py
@@ -151,24 +151,26 @@ def test_context_inerror_rollback_no_clobber(conn, dsn, caplog):
     assert "in rollback" in rec.message
 
 
-def test_context_active_rollback_no_clobber(conn, dsn, caplog):
+def test_context_active_rollback_no_clobber(dsn, caplog):
     caplog.set_level(logging.WARNING, logger="psycopg")
 
-    with pytest.raises(ZeroDivisionError):
-        conn2 = Connection.connect(dsn)
-        with conn2.transaction():
-            with conn2.cursor() as cur:
-                with cur.copy(
-                    "copy (select generate_series(1, 10)) to stdout"
-                ) as copy:
-                    for row in copy.rows():
-                        1 / 0
+    conn = Connection.connect(dsn)
+    try:
+        with pytest.raises(ZeroDivisionError):
+            with conn.transaction():
+                conn.pgconn.exec_(
+                    b"copy (select generate_series(1, 10)) to stdout"
+                )
+                status = conn.info.transaction_status
+                assert status == conn.TransactionStatus.ACTIVE
+                1 / 0
 
-    assert len(caplog.records) == 1
-    rec = caplog.records[0]
-    assert rec.levelno == logging.WARNING
-    assert "in rollback" in rec.message
-    conn2.close()
+        assert len(caplog.records) == 1
+        rec = caplog.records[0]
+        assert rec.levelno == logging.WARNING
+        assert "in rollback" in rec.message
+    finally:
+        conn.close()
 
 
 def test_interaction_dbapi_transaction(conn):